update conanfile.py to check if the large data files were pulled from the repo.
Even though the BSK install instructions say to install lfs first before cloning the repo, this is a
common source of errors with new BSK users.  We are adding a check to the
conanfile.py script to ensure that de430.bsp is a large file.  Otherwise the script ends
with an error messaging asking the user to install lfs first before cloning.